+++ project/build.scala
@@ -14,11 +14,10 @@ object PrimesscalatraappBuild extends Build {
 
   val Organization    = "fr.janalyse"
   val Name            = "primesui"
-  val Version         = penvOrElse("PRIMESUI_REV", "0.1.33-SNAPSHOT")
+  val Version         = penvOrElse("PRIMESUI_REV", "0.1.34-SNAPSHOT")
   val PrimesVersion   = penvOrElse("PRIMES_REV",   "1.2.2-SNAPSHOT")
   val ScalaVersion    = "2.11.7"
-  val ScalatraVersion = "2.4.0.RC3"
-  //val ScalatraVersion = "2.4.0-SNAPSHOT"
+  val ScalatraVersion = "2.4.0"
 
   lazy val project = Project(
     "primes-scalatra-app",